stateful selectbox + pinpoint persistence
